refactor(insights): rename and split property math selector components
